# specs/005-/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Crystal Intelligence PoC API
  version: 0.1.0
  description: >-
    Streamlit UIとMCPサーバーから利用されるNeo4jベースのナレッジAPI。
    すべてのエンドポイントはPython 3.11 + FastAPIを想定し、OpenAI APIキーはヘッダーで受け渡される。
servers:
  - url: http://localhost:8000
    description: Local development (Docker Compose)
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
  /api/query:
    post:
      summary: Run agent against user query and persist knowledge
      operationId: postQuery
      tags: [agent]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        '200':
          description: Query executed and knowledge persisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Agent failure (all retries exhausted)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/knowledge/search:
    get:
      summary: Search personal + team knowledge with vector ranking
      operationId: searchKnowledge
      tags: [search]
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: user_id
          required: true
          schema:
            type: string
        - in: query
          name: top_k
          required: false
          schema:
            type: integer
            default: 100
            maximum: 100
        - in: query
          name: include_team
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Search results with provenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/batch/promote:
    post:
      summary: Run promotion batch to elevate personal knowledge to team level
      operationId: promoteKnowledge
      tags: [batch]
      requestBody:
        required: false
        description: Optional overrides for similarity threshold or dry-run mode
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionRequest'
      responses:
        '200':
          description: Promotion job executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionResponse'
        '400':
          description: Invalid configuration or missing prerequisites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /healthz:
    get:
      summary: Health probe for UI/CI
      operationId: healthz
      tags: [ops]
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                  neo4j:
                    type: string
                  openai:
                    type: string
        '503':
          description: Dependency failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    QueryRequest:
      type: object
      required: [user_id, query]
      properties:
        user_id:
          type: string
        query:
          type: string
          minLength: 1
        channels:
          type: array
          description: Agent channels to enable (tools)
          items:
            type: string
            enum: [file_search, web_search, query_team_knowledge]
          default: []
        metadata:
          type: object
          additionalProperties:
            type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    Attachment:
      type: object
      required: [path, mime_type]
      properties:
        path:
          type: string
        mime_type:
          type: string
        checksum:
          type: string
    QueryResponse:
      type: object
      required: [answer, knowledge_id, knowledge_level, provenance, references, latency_ms]
      properties:
        answer:
          type: string
        knowledge_id:
          type: string
        knowledge_level:
          type: string
          enum: [personal, team]
        provenance:
          type: array
          items:
            $ref: '#/components/schemas/ProvenanceItem'
        references:
          type: array
          items:
            $ref: '#/components/schemas/ProvenanceItem'
        logs_written:
          type: boolean
        retries:
          type: integer
        latency_ms:
          type: number
          format: float
        tool_traces:
          type: array
          items:
            $ref: '#/components/schemas/ToolTrace'
    ProvenanceItem:
      type: object
      required: [id, source_type]
      properties:
        id:
          type: string
        source_type:
          type: string
          enum: [file, web, knowledge, agent_synthesis]
        summary:
          type: string
        similarity:
          type: number
          format: float
        owner:
          type: object
          properties:
            user_id:
              type: string
            team_id:
              type: string
        created_at:
          type: string
          format: date-time
    SearchResponse:
      type: object
      required: [query, results]
      properties:
        query:
          type: string
        latency_ms:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchHit'
    SearchHit:
      type: object
      required: [knowledge_id, rank, similarity, level]
      properties:
        knowledge_id:
          type: string
        level:
          type: string
          enum: [personal, team]
        title:
          type: string
        summary:
          type: string
        similarity:
          type: number
          format: float
        rank:
          type: integer
        owner_label:
          type: string
        created_at:
          type: string
          format: date-time
        provenance:
          type: array
          items:
            $ref: '#/components/schemas/ProvenanceItem'
    PromotionRequest:
      type: object
      properties:
        dry_run:
          type: boolean
          default: false
        similarity_threshold_override:
          type: number
          minimum: 0.0
          maximum: 1.0
        min_contributors_override:
          type: integer
          minimum: 1
        min_team_ratio_override:
          type: number
          minimum: 0.0
          maximum: 1.0
    PromotionResponse:
      type: object
      required: [summary, promoted, skipped]
      properties:
        summary:
          type: object
          properties:
            total_candidates:
              type: integer
            promoted_count:
              type: integer
            skipped_count:
              type: integer
        promoted:
          type: array
          items:
            $ref: '#/components/schemas/PromotionResult'
        skipped:
          type: array
          items:
            $ref: '#/components/schemas/PromotionResult'
    PromotionResult:
      type: object
      required: [group_id, status]
      properties:
        group_id:
          type: string
        status:
          type: string
          enum: [promoted, skipped]
        reasons:
          type: array
          items:
            type: string
        similarity:
          type: number
        contributors:
          type: array
          items:
            type: string
        team_coverage_ratio:
          type: number
          format: float
        knowledge_id:
          type: string
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ToolTrace:
      type: object
      required: [name, input, output]
      properties:
        name:
          type: string
        input:
          type: object
          additionalProperties: true
        output:
          type: object
          additionalProperties: true